+++ packages/experimental/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @effect/experimental
 
+## 0.16.2
+
+### Patch Changes
+
+- Updated dependencies [[`1f6dc96`](https://github.com/Effect-TS/effect/commit/1f6dc96f51c7bb9c8d11415358308604ba7c7c8e), [`1f6dc96`](https://github.com/Effect-TS/effect/commit/1f6dc96f51c7bb9c8d11415358308604ba7c7c8e)]:
+  - @effect/platform-node@0.49.2
+  - effect@3.1.3
+  - @effect/platform@0.53.2
+  - @effect/schema@0.66.16
+
 ## 0.16.1
 
 ### Patch Changes