+++ packages/platform-node-shared/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @effect/platform-node-shared
 
+## 0.4.21
+
+### Patch Changes
+
+- Updated dependencies [[`1f6dc96`](https://github.com/Effect-TS/effect/commit/1f6dc96f51c7bb9c8d11415358308604ba7c7c8e)]:
+  - effect@3.1.3
+  - @effect/platform@0.53.2
+
 ## 0.4.20
 
 ### Patch Changes