+++ packages/printer/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @effect/printer
 
+## 0.33.12
+
+### Patch Changes
+
+- Updated dependencies [[`1f6dc96`](https://github.com/Effect-TS/effect/commit/1f6dc96f51c7bb9c8d11415358308604ba7c7c8e)]:
+  - effect@3.1.3
+  - @effect/typeclass@0.24.12
+
 ## 0.33.11
 
 ### Patch Changes