+++ packages/sql-sqlite-react-native/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @effect/sql-sqlite-react-native
 
+## 0.3.4
+
+### Patch Changes
+
+- Updated dependencies [[`1f6dc96`](https://github.com/Effect-TS/effect/commit/1f6dc96f51c7bb9c8d11415358308604ba7c7c8e)]:
+  - effect@3.1.3
+  - @effect/sql@0.2.4
+
 ## 0.3.3
 
 ### Patch Changes